+++ RpiMotorLib/RpiMotorScriptLib.py
@@ -15,7 +15,7 @@
 import argparse
 import sys
 
-__version__ = "2.2-3"
+__version__ = "2.3-4"
 __author__ = "Gavin Lyons"
 __url__ = "https://github.com/gavinlyonsrepo/RpiMotorLib"