+++ RpiMotorLib/rpi_dc_lib.py
@@ -13,13 +13,15 @@
 # ========================== IMPORTS ======================
 # Import the system modules needed to run rpiMotorlib.py
 import time
-import RPi.GPIO as GPIO 
+import RPi.GPIO as GPIO
 
 # ==================== CLASS SECTION ===============================
 
 
 class L298NMDc(object):
-    """ Class to control DC motor via L298n motor controller """
+    """ Class to control DC motor via L298n motor controller
+    6 methods 1. __init__ 2. forward
+    3.backward 4.stop 5 .brake 6.cleanup"""
 
     def __init__(self, pin_one, pin_two,
                  pwm_pin, freq=50, verbose=False, name="DCMotorX"):
@@ -89,8 +91,9 @@ class L298NMDc(object):
         """ brake motor"""
         GPIO.output(self.pin_one, True)
         GPIO.output(self.pin_two, True)
-        if duty_cycle != self.last_pwm:
+        if self.verbose:
             print("Braking Motor : Duty Cycle = {}".format(duty_cycle))
+        if duty_cycle != self.last_pwm:
             self.my_pwm.ChangeDutyCycle(duty_cycle)
             self.last_pwm = duty_cycle
 
@@ -103,6 +106,90 @@ class L298NMDc(object):
         self.my_pwm.ChangeDutyCycle(0)
         if clean_up:
             GPIO.cleanup()
+
+
+class DRV8833NmDc(object):
+    """ Class to control DC motor via L298n motor controller
+    6 methods 1. __init__ 2. forward
+    3.backward 4.stop 5.brake 6.cleanup"""
+
+    def __init__(self, pin_one, pin_two,
+                 freq=50, verbose=False, name="DCMotorY"):
+        """ init method
+        (1) pin_one, type=int,  GPIO pin  direction pin connected to IN1 or IN3
+        (2) Pin two type=int, GPIO pin PWM speed pin connected to IN2 or IN4
+        (3) freq in Hz default 50
+        (4) verbose, type=bool  type=bool default=False
+         help="Write pin actions"
+        (5) name, type=string, name attribute
+        """
+        self.name = name
+        self.pin_one = pin_one
+        self.pin_two = pin_two
+        self.freq = freq
+        self.verbose = verbose
+
+        GPIO.setmode(GPIO.BCM)
+        GPIO.setwarnings(False)
+        GPIO.setup(self.pin_one, GPIO.OUT)
+        GPIO.setup(self.pin_two, GPIO.OUT)
+
+        self.my_pwm = GPIO.PWM(self.pin_two, self.freq)
+        self.last_pwm = 0
+        self.my_pwm.start(self.last_pwm)
+        if self.verbose:
+            print(" Motor initialized named: {} ".format(self.name))
+            print(" Direction pin In1 or In3:  {}".format(self.pin_one))
+            print(" PWM speed pin In2 or in4:  {}".format(self.pin_two))
+            print(" Frequency: {} ".format(self.freq))
+
+    def forward(self, duty_cycle=50):
+        """ Move motor forwards passed duty cycle for speed control """
+        GPIO.output(self.pin_one, True)
+        if self.verbose:
+            print("Moving Motor Forward : Duty Cycle = {}".format(duty_cycle))
+        if duty_cycle != self.last_pwm:
+            self.my_pwm.ChangeDutyCycle(duty_cycle)
+            self.last_pwm = duty_cycle
+
+    def backward(self, duty_cycle=50):
+        """ Move motor backwards passed duty cycle for speed control"""
+        GPIO.output(self.pin_one, False)
+        if self.verbose:
+            print("Moving Motor Backward : Duty Cycle = {}".format(duty_cycle))
+        if duty_cycle != self.last_pwm:
+            self.my_pwm.ChangeDutyCycle(duty_cycle)
+            self.last_pwm = duty_cycle
+
+    def stop(self, duty_cycle=0):
+        """ Stop motor"""
+        GPIO.output(self.pin_one, False)
+        self.my_pwm.ChangeDutyCycle(duty_cycle)
+        if self.verbose:
+            print("Stoping Motor : Duty Cycle = {}".format(duty_cycle))
+        if duty_cycle != self.last_pwm:
+            self.my_pwm.ChangeDutyCycle(duty_cycle)
+            self.last_pwm = duty_cycle
+                
+    def brake(self, duty_cycle=100):
+        """ brake motor"""
+        GPIO.output(self.pin_one, True)
+        if self.verbose:
+            print("Braking Motor : Duty Cycle = {}".format(duty_cycle))
+        if duty_cycle != self.last_pwm:
+            self.my_pwm.ChangeDutyCycle(duty_cycle)
+            self.last_pwm = duty_cycle
+       
+    def cleanup(self, clean_up=False):
+        """ cleanup all GPIO connections used in event of error by lib user"""
+        if self.verbose:
+            print("rpi_dc_lib.py : Cleaning up")
+        GPIO.output(self.pin_one, False)
+        self.my_pwm.ChangeDutyCycle(0)
+        GPIO.output(self.pin_two, False)
+        if clean_up:
+            GPIO.cleanup()
+
 
 class TranDc(object):
     """ Class to control DC motor via a transistor """
@@ -132,12 +219,12 @@ class TranDc(object):
         if self.verbose:
             print("Speed PWM duty cycle percentage {}".format(speed))
 
-
     def dc_clean_up(self):
         """ docstring """
         self.motor_pwm.ChangeDutyCycle(0)
         self.motor_pwm.stop()
         GPIO.output(self.pin, False)
+
 
 def importtest(text):
     """import print test statement"""